fix(inf): fix account loading in infra mode
Account kubeconfig files were not being mounted correctly under the
previous config format. This commit introduces the ability to mount
arbitrary numbers of kubeconfig files passed in via invocation or
values.yml files. These files are then loaded into an opaque secret and
given a unique name relative to the helm chart. This secret will then be
mounted into the agent container and referenced by name in the agent
config file.

Additionally, this PR removes the service account and role created when
running in infrastructure mode. IOW, this change makes it mutually
exclusive to run in either agent more OR infrastructure mode, but not
both.